+++ CHANGELOG.md
@@ -8,6 +8,20 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 <!-- next-header -->
 
 ## [Unreleased] - ReleaseDate
+- Fix breadcrumbs for right-to-left languages [#489](https://github.com/svenstaro/miniserve/pull/489) (thanks @aliemjay)
+- Fix URL percent encoding for special characters [#485](https://github.com/svenstaro/miniserve/pull/485) (thanks @aliemjay)
+- Wrap breadcrumbs at any char [#496](https://github.com/svenstaro/miniserve/pull/496) (thanks @aliemjay)
+- Add separate flags for compressed and uncompressed tar archives [#492](https://github.com/svenstaro/miniserve/pull/492) (thanks @deantvv)
+
+## [0.13.0] - 2021-03-28
+- Change default log level to `Warn`
+- Change some messages a bit to be more clear
+- Add `--print-completions` to print shell completions for various supported shells [#482](https://github.com/svenstaro/miniserve/pull/482) (thanks @rouge8)
+- Don't print some messages if not attached to an interactive terminal
+- Refuse to start if not attached to interactive terminal and no explicit path is provided
+
+  This is a security consideration as you wouldn't want to run miniserve without an explicit path
+  as a service. You could end up serving `/` or `/root` in case those working directories are set.
 
 ## [0.12.1] - 2021-03-27
 - Fix QR code not showing when using both `--random-route` and `--qrcode` [#480](https://github.com/svenstaro/miniserve/pull/480) (thanks @rouge8)
@@ -73,7 +87,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Some theme related bug fixes (thanks @boastful-squirrel)
 
 <!-- next-url -->
-[Unreleased]: https://github.com/svenstaro/miniserve/compare/v0.12.1...HEAD
+[Unreleased]: https://github.com/svenstaro/miniserve/compare/v0.13.0...HEAD
+[0.13.0]: https://github.com/svenstaro/miniserve/compare/v0.12.1...v0.13.0
 [0.12.1]: https://github.com/svenstaro/miniserve/compare/v0.12.0...v0.12.1
 [0.12.0]: https://github.com/svenstaro/miniserve/compare/v0.11.0...v0.12.0
 [0.11.0]: https://github.com/svenstaro/miniserve/compare/v0.10.4...v0.11.0